Move exceptions to shares_base
