+++ backend/dataall/modules/s3_datasets_shares/services/share_item_service.py
@@ -21,7 +21,7 @@ from dataall.modules.shares_base.db.share_object_state_machines import (
     ShareObjectSM,
     ShareItemSM,
 )
-from dataall.modules.s3_datasets_shares.services.share_exceptions import ShareItemsFound
+from dataall.modules.shares_base.services.share_exceptions import ShareItemsFound
 from dataall.modules.s3_datasets_shares.services.share_notification_service import ShareNotificationService
 from dataall.modules.shares_base.services.share_permissions import (
     GET_SHARE_OBJECT,
