+++ backend/dataall/modules/s3_datasets_shares/services/share_managers/s3_bucket_share_manager.py
@@ -13,7 +13,7 @@ from dataall.modules.s3_datasets_shares.aws.kms_client import (
 )
 from dataall.modules.s3_datasets_shares.aws.s3_client import S3ControlClient, S3Client, DATAALL_READ_ONLY_SID
 from dataall.modules.shares_base.db.share_object_models import ShareObject
-from dataall.modules.s3_datasets_shares.services.share_exceptions import PrincipalRoleNotFound
+from dataall.modules.shares_base.services.share_exceptions import PrincipalRoleNotFound
 from dataall.modules.s3_datasets_shares.services.share_managers.share_manager_utils import ShareErrorFormatter
 from dataall.modules.s3_datasets_shares.services.dataset_sharing_alarm_service import DatasetSharingAlarmService
 from dataall.modules.s3_datasets_shares.services.managed_share_policy_service import (
