+++ backend/dataall/modules/s3_datasets_shares/services/share_object_service.py
@@ -26,7 +26,7 @@ from dataall.modules.shares_base.db.share_object_state_machines import (
     ShareObjectSM,
     ShareItemSM,
 )
-from dataall.modules.s3_datasets_shares.services.share_exceptions import ShareItemsFound, PrincipalRoleNotFound
+from dataall.modules.shares_base.services.share_exceptions import ShareItemsFound, PrincipalRoleNotFound
 from dataall.modules.s3_datasets_shares.services.share_item_service import ShareItemService
 from dataall.modules.s3_datasets_shares.services.share_notification_service import ShareNotificationService
 from dataall.modules.s3_datasets_shares.services.managed_share_policy_service import SharePolicyService
