+++ backend/dataall/modules/shares_base/services/share_exceptions.py
